+++ packages/plugin-mpa/src/index.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import { getMpaEntries } from '@builder/app-helpers';
+import { getMpaEntries, formatPath } from '@builder/app-helpers';
 import { generateMPAEntries } from '@builder/mpa-config';
 import { IPlugin } from 'build-scripts';
 
@@ -100,15 +100,17 @@ const plugin: IPlugin = (api) => {
     }
     // modify entry
     modifyUserConfig('entry', finalMPAEntries);
-    applyMethod('addImportDeclaration', {
-      multipleSource: {
-        runApp: redirectEntries.map(({ entryPath, runAppPath }) => ({
-          filename: entryPath,
-          value: runAppPath,
-          type: 'normal',
-        })),
-      },
-    });
+    if (redirectEntries.length > 0) {
+      applyMethod('addImportDeclaration', {
+        multipleSource: {
+          runApp: redirectEntries.map(({ entryPath, runAppPath }) => ({
+            filename: entryPath,
+            value: formatPath(runAppPath),
+            type: 'normal',
+          })),
+        },
+      });
+    }
     // set page template
     onGetWebpackConfig(config => {
       setPageTemplate(rootDir, entries, (mpa as any).template || {}, config, setValue);