+++ packages/plugin-router/src/runtime.tsx
@@ -1,56 +1,20 @@
 import * as React from 'react';
 // @ts-ignore
-import ErrorBoundary from '$ice/ErrorBoundary';
-// @ts-ignore
 import defaultRoutes from '$ice/routes';
 import { IceRouter, Routes, parseRoutes } from './runtime/Router';
-import { wrapperPageWithCSR, wrapperPageWithSSR } from './runtime/formatRoutes';
 import { RouteItemProps } from './types/base';
 import { IRouterConfig } from './types';
 
-const module = ({ setRenderApp, appConfig, modifyRoutes, wrapperPageComponent, modifyRoutesComponent, buildConfig, context, applyRuntimeAPI }) => {
-  const { router: appConfigRouter = {}, app = {} } = appConfig;
-  const { ErrorBoundaryFallback, onErrorBoundaryHandler } = app;
+const module = ({ setRenderApp, appConfig, modifyRoutes, modifyRoutesComponent, buildConfig, context, applyRuntimeAPI }) => {
+  const { router: appConfigRouter = {} } = appConfig;
 
-  const { parseSearchParams = true } = app;
-  const WrappedPageComponent = (PageComponent) => {
-    const InnerWrappedPageComponent = (props) => {
-      const searchParams = parseSearchParams && applyRuntimeAPI('getSearchParams');
-      return <PageComponent {...Object.assign({}, props, { searchParams })} />;
-    };
-    return InnerWrappedPageComponent;
-  };
-
-  wrapperPageComponent(WrappedPageComponent);
-
-  // plugin-router 内置确保了 defaultRoutes 最先被添加
   modifyRoutes(() => {
-    // TODO: format routes in the src/runtime/Router.tsx
-    // return formatRoutes(appConfigRouter.routes || defaultRoutes, '');
     return appConfigRouter.routes || defaultRoutes;
   });
 
   // add default RoutesComponent
   modifyRoutesComponent(() => Routes);
 
-  const wrapperPageErrorBoundary = (PageComponent) => {
-    const { pageConfig = {} } = PageComponent;
-    const WrappedPageErrorBoundary = (props) => {
-      if (pageConfig.errorBoundary) {
-        return (
-          <ErrorBoundary Fallback={ErrorBoundaryFallback} onError={onErrorBoundaryHandler}>
-            <PageComponent {...props} />
-          </ErrorBoundary>
-        );
-      }
-      return <PageComponent {...props} />;
-    };
-    return WrappedPageErrorBoundary;
-  };
-
-  const wrapperPageFn = process.env.__IS_SERVER__ ? wrapperPageWithSSR(context) : wrapperPageWithCSR();
-  wrapperPageComponent(wrapperPageFn);
-  wrapperPageComponent(wrapperPageErrorBoundary);
   if (appConfigRouter.modifyRoutes) {
     modifyRoutes(appConfigRouter.modifyRoutes);
   }