+++ packages/build-app-templates/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- [chore] use `enableRouter` replace `buildConfig.router`
+- [feat] add `runtimeValue.enableRouter` when createBaseApp
+
 ## 1.0.2
 
 - [fix] add default static config of rax app render