+++ packages/build-user-config/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 2.0.1
+
+- [fix] enhance alias for node_modules dependencies
+
+## 2.0.0
+
+- [fix] add url fallback for webpackHotDevClient
+- [fix] bump version of `@builder/pack`(^0.5.0)
+- [feat] add https url list
+- [chore] enhance `applyCliOptions`
+- [chore] RedirectPathLoader should include `.rax|.ice`
+- [fix] skip deps `filepath` in windows
+
+## 1.1.9
+
+- [feat] compact compact postcss options in different version of postcss-loader
+
+## 1.1.8
+
+- [fix] re-calculate webpack cache id when config `disableRuntime`
+
 ## 1.1.7
 
 - [fix] undefined `postcssOptions.plugins`
@@ -39,7 +60,6 @@
 - [feat] refactor config of eslint
 - [feat] support active minify in dev
 
-
 ## 1.0.2
 
 - [fix] add dependency of eslint-reporting-webpack-plugin
@@ -78,7 +98,7 @@
 
 ## 0.3.7
 
-- [feat] add `devServer.host` to https cert 
+- [feat] add `devServer.host` to https cert
 - [chore] `tabItem.path` => `tabItem.pageName`/ `tabItem.name` => `tabItem.text`
 - [fix] compatible with no loader specified when using `postcssrc: true` in ssr app