+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 3.0.6
+
+- [fix] server-side render failed in production(always render `global.__ICE_SERVER_HTML_TEMPLATE__`)
+- [chore] remove `build/server/loadable-stats.json`
+
+## 3.0.5
+
+- [fix] update dependencies version
+- [feat] support no router mode ssr
+
+## 3.0.4
+
+- [fix] compatible with special charts in html content
+
 ## 3.0.3
 
 - [fix] route path value may be undefined