+++ packages/webpack-dev-mock/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.0
+
+- [fix] bump version of `@builder/babel-config`(^2.0.0)
+
+## 1.3.2
+
+- [fix] match the correct pathname of request
+
 ## 1.3.1
 
 - [fix] update dependencies